requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dlib==19.22.0
Flask==1.1.2
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
numpy==1.20.2
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
scipy==1.6.2
six==1.15.0
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
waitress==2.0.0